requirements.txt
openai-agents>=0.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
email-validator>=2.0.0
aiofiles>=23.0.0
openpyxl>=3.1.0
